[FIX] mail_enterprise: display all attachements in preview widget
The attachment preview widget used to display only attachments posted in the
mail threads, and not the attachments directly linked to the record.
This means that only the files that are put into the mail thread
could be used for the OCR.

To sum up, the renderer has a chatter widget that contains the correct list of
attachments, obtained through an RPC.
It also has a thread_widget which only contains a subset of attachments,
those related to messages.
It is the ThreadField that triggers up the event to render the AttachmentPreview
which is caught by the formRenderer, with its list of attachments.
Instead we don't pass the attachments and we intercept the event in the chatter.
If the chatter already has the attachments loaded,
we add the list to the event data, and let it bubble up.
Otherwise we stop the event, fetch the list, then trigger it again.

opw 1950403

closes odoo/enterprise#3846

Signed-off-by: Nans Lefebvre (len) <len@odoo.com>