[FIX] l10n_ar_reports: fixed typo and sign of formula in vat report
report line 'Impuesto Cobrado IVA' should hold tax amount
of sales invoices/refund instead of vendor bills/refund

bug was spotted while migrating this module
to V13 tax configuration task

By checking with Adhoc, also either the sign needed to change
or the formula as purchases need to be subtracted from sales.
We chose to change the formula for both the base imponible
and the impuestos.  (base amount and tax)

Closes #3860

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>