[FIX] l10n_mx_report: Allow to show the journal items in the parent lines
In the trial balance was added the option to open the journal items when
the lines has the next estructure:

`line.get('id') != 'hierarchy' and line.get('unfolded')`

And in the Mexican report, some lines do not has the account reference
(lines with tag relation to allow group). Then try to show an account
that is incorrect in that cases and show other account or the message:

`Record does not exist or has been deleted.`

Now search in the last level, to show all the journal items in the
period for the accounts in the group selected.