[FIX] web_studio: modif next to comment
It is possible to get in a situation where a modification will be just
next to an XML comment.

Before 858638f13 this was not an issue since in is_valid_anchor, the
gotten invalid xpath:
 .//data[1]/<cyfunction Comment at 0x7f32d3b57830>[1]

(it is like this since comment_node.tag is a function and not a string)

would just be ignored by Element.find method.

But since 858638f13:

- we expect the .tag to be a string
- Element.xpath instead of .find do not allow invalid xpaths

opw-1957048
closes #3872

Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>