[FIX] mrp_workorder: show close MO button only when no other WO is left to process
Suppose that on the routing, the operation uses batch and batch_size.
Then multiple WOs can be processed in parallel.
However it was always assumed that all previouses workorders had been entirely
processed before the current one.
As a result of this assumption, the "close MO" button would appear when
processing the last workorder, even if previous ones were still processing.

opw 1943674

closes odoo/enterprise#3815

Signed-off-by: Nans Lefebvre (len) <len@odoo.com>