[FIX] sale_ebay: put ebay synchronized fields in copy=False
Essentially, a product is synchornized to ebay through an ebay listing,
with a bunch of parameters (ebay title, description, etc...) that are used for
the listing.
However that means that there are essentially two types of fields, the ones that
are used to describe the listing and 'Odoo private ones' that are sued to manage
the listing.
Fields of the latter kind are e.g. ebay_last_sync, ebay_id or ebay_url.
However since there are many fields of the former kind,
it is useful to duplicate a product to have them prefilled.

What happens if the fields of the second kind are not in copy=False is that the
two products are synchronized to the same product on ebay which can cause
further bugs down the line
(e.g. transactions not synchronized because of an incorrect ebay_last_sync).

opw 1924454

closes odoo/enterprise#3884

Signed-off-by: Nans Lefebvre (len) <len@odoo.com>