[IMP] l10n_mx_edi: Do not allow to delete a payment signed in the SAT
In the next case:
1. Generate an Mexican invoice and sign
2. Make the invoice payment with a bank statement and generate the
payment complement.
3. Un-reconcile the bank statement.

In this moment Odoo, remove the move in the payment generated from the
bank statement and unlink the record without cancel, then the CFDI is
not cancelled in the SAT system.

Now is called the cancel method when is unlink the payment and if it
cannot be cancelled show an raise error because the document is valid in
the SAT.

closes odoo/enterprise#3894

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>