[FIX] l10n_mx_edi: payment with mark invoice fully paid
Have a customer invoice of 1000 MXN, register a payment of 999.99 MXN,
Mark invoice as fully paid.

Open the generated CFDI Comprobante file.

Before this commit, In the line DoctoRelacionado, where the information
the payment by invoices is shown, the field ImpPagado is 1000 and
not 999.99

Now, the corresponding amount of the payment for the invoice is shown.

For more information see : http://omawww.sat.gob.mx/tramitesyservicios/Paginas/documentos/Guia_comple_pagos.pdf

opw-1952963

closes odoo/enterprise#3935

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>