[FIX] web_enterprise: open app in one click on IE
On internet explorer 11, we had to click two times on an app on the
dashboard to be able to open it.

This is because on IE, an element with placeholder and no value trigger:

- an input event after a focus event on itself
- an input event before a blur event on itself

So on IE when the dashboard is opened, we:

- focus on the search input (=> ok for all browser)
- dashboard gets in searching mode because input event (=> not ok)
- [the user click on an app in the dashboard]
- the placeholder content is added in the search input
- there is a input event in search input
- the dashboard is updated and rerendered
- the click did not work (=> not ok)
- [the user click a second time on the app in the dashboard]
- the click works

On other browsers, the dashboard is not in searching mode at the start,
and only one click is necessary to go to an application.

With this changeset, search input events are ignored if we were not in
searching mode and the search value is empty (because then this probably
is an erroneous placeholder input event).

This is for odoo [10.0,11.0], in saas-11.3 home menu refacturing most of
the issue is removed (since there is no autofocus anymore on search
input).

jsfiddle of the IE11 issue: https://jsfiddle.net/093vzw7k/1/ when
focusing then blurring the first field we get:
- on other browsers: bBfF (focus then blur)
- on IE11: biBIifIF (focus then input then input then blur)

opw-1959340
closes #3936

Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>