[FIX] account_bank_statement_import_camt: do not use transaction origin currency when not in multicurrency
Commits de7c2a97 and 57ab4c59 made it so that when importing transactions
in a CAMT xml file, the transaction currency C1 is correctly captured;
this currency may differ from the entry currency,
which is C2, the currencty of the target bank account.
Independently, a vendor bill might be in currency C3.
Ideally, to make the reconciliation easy we want to have C1=C2=C3,
but in the general case they might all be different from other.
Note that in particular C1 is the customer's currency,
so the user has no control over it.

However if the user is not in multicurrency, we know that C2=C3.
So in that case it is better to ignore C1 and just keep C2,
since it makes the amounts match.
Note that this is made problematic by the fact that without multicurrency,
it is likely that all currency rates are still set to their default value (1),
so all amounts conversions appear to be wrong.

opw 1943400

closes odoo/enterprise#3988

Signed-off-by: Nans Lefebvre (len) <len@odoo.com>