[FIX] marketing_automation: utm info on sent mail
Oversight of odoo/odoo#29083, the UTM data of mass mailing were not
overrided by the marketing compaign.

Before odoo/odoo#29083 the campaign and source were respectively the
marketing campaign and marketing activity. After it was data about the
mass mailing which can be common between several marketing activities.

With this changeset, UTM data is same as before odoo/odoo#29083.

opw-1903803
closes #3996

Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>