[FIX] l10n_mx_edi_sale_coupon: negative lines in invoices with coupons
cherry-pick of 29930f4a249a763bdcb932668dd25cf2f5e4c724

Until now, in Odoo, when a SO with a coupon applied generated an invoice,
a negative line with the discount's amount (e.g., 10% of the invoice)
was added to the invoice. This generates an issue with the Mexican CFDI
(Comprobante Fiscal Digital por Internet) generation, in which the
'Valor Unitario' (unit price) of each line of the invoice, must be
strictly positive (see pp 20 of
http://omawww.sat.gob.mx/informacion_fiscal/factura_electronica/Documents/Gu%C3%ADaAnexo20.pdf).

The same document, in pp 47, explain that for the discounts, the
reductions and the bonuses a credit note (comprobante de egreso) must be
generated.

Now, when a SO with a coupon applied generates not one invoice, but two
invoices: one customer invoice and one related credit note.
In the customer invoice, we will find the invoice without the discount
and in the credit note we will find the discount's amount. This will
generate two related Mexican CFDI (Comprobante de Ingreso and Comprobante
de Egreso).

opw-1959385
opw-1962939

closes odoo/enterprise#4355

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>