[FIX] account_sepa_direct_debit: multi invoices
- Create a mandate for a partner A
- Create 2 invoices, one for A and one for another partner without
  mandate
- Select both invoices and choose 'Confirm Draft Invoices'

An error occurs: 'You can only register payments for open invoices'.

The payment is supposed to be created in `pay_with_mandate` for the
invoice of partner A. However, `active_ids` are retrieved in the
`default_get` method of the `account.payment`:

https://github.com/odoo/odoo/blob/12.0/addons/account/models/account_payment.py#L75

We should clean-up the context in this case.

opw-1962424

closes odoo/enterprise#4009

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>