[FIX] web_studio: nondeterministic on runbot
The issue here was in the new report tour. It requires to create a new
record to correctly display the report but Studio was opened before the
record creation so it triggers a modal with 'The record has been
modified, ...' which prevented the tour to continue.

An previous attempt to fix this bug has been done in odoo/enterprise@59d90dd60
with an `extra_trigger` but without a great success.

The record creation at the beginning of the tour is in fact useless
since rev. odoo/odoo@a06e6ca because the report tour is now directly
executed after the 'New App' tour, in which a record has already been
created.

Note that the two first steps (toggling Studio and opening the app) have
been inverted to avoid another issue (try to toggle Studio while opening
the app, which is asynchronous).

Task 1937583

closes odoo/enterprise#4015

Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>