[FIX] account_invoice_extract: fix show OCR button dependence
The _compute_show_send/resend_button methods did not actually depend
on message_ids, but on message_main_attachment_id.
As a side-effect, the two were essentially the same as long as only attachments
in the thread were used.
Since the series of commit to use all the attachments linked to the record,
it makes sense to actually use the correct dependence
(since it can be modified at attachment upload and not only at message_post).

opw 1950403