[FIX] web_enterprise: prevent hash change if not needed
This commit fixes a rare race condition. It is very unlikely to happen
to a real user, but the problem happens frequently on automated tests in
some cases.

We noticed the problem with the tour rte: it sometimes failed at the
first step.  The various steps are:
1. python test starts a chrome process on /web
2. whenever it can, it starts the rte tour
3. the tour manager sees that the rte tour is on a different url, and
redirects the chrome browser to that url (by setting its location.href)
4. this initiates a GET request, which is obviously asynchronous
5. the normal js from the web browser keeps working, because the GET
request is not completed
6. at some point, the web client toggle its home menu, which add a #home
at the end of the url by setting its location.href
7. then the GET request completes, and is dropped silently because we
are in a different location.

The underlying issue is that the code that triggers the hash change has
no idea that there is a pending request.  And without some specialized
code (for example, rerouting all navigation changes through some code
that queues them), we have no real way of knowing that.

So, what we do in this commit is to avoid this specific situation entirely
by simply not setting the hash to #home initially. This is not a general
fix, but since this is the only occurence of this issue that I am aware
of, this should be enough.

closes odoo/enterprise#4019

Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>

Backport of 3d9d117790fe0fc8c49ac84f4e33f00e2aeda9c1

closes odoo/enterprise#4356

Signed-off-by: Christophe Simonis <chs@odoo.com>