[FIX] account_invoice_extract, mail_enterprise: synchronize the main_message_attachment
The main_message_attachment_id is the file that is sent through OCR.
Without this change, the preview is not synchronized with the main,
unless an explicit call to switch main is made.

The information was actually put in _message_read_dict_postprocess,
but this only contains the messages from the mail thread.
Since the chatter messages are obtained through a normal search_read,
we can't bake that information in the server return.

opw 1950403

closes odoo/enterprise#4002

Signed-off-by: Nans Lefebvre (len) <len@odoo.com>