[FIX] web_enterprise: properly relayout menu in some cases
The issue solved by this commit is the following:
1. open an odoo application with a large number of menu items
2. open the home menu (this hides the menu items)
3. resize the window application to reduce the window size
4. close the home menu
5. The menuitems are not computed are are messed up

With this fix, we force a relayout computation each time the user menu
is retoggled. This is not optimal, but this is necessary since the
relayout code is actually in a separate file in community (dom.js), and
not accessible from the outside.

opw 1965432

closes odoo/enterprise#4045

Signed-off-by: Géry Debongnie (ged) <ged@openerp.com>