[FIX] delivery_usps: WeightInOunces in ounces
Thanks to odoo/odoo@52a85e57b and b8f3b8ccae as of saas-11.3 it is
possible to choose between Pound and Kilogram in general settings
"Weight unit of measure" setting.

Before the default was only in kilogram.

There was a small part missing for USPS carrier that is fixed with this
change, with it if customer is in "Pound" default, the WeightInOunces
will not be converted from kilogram to pounds.

opw-1959463
closes #4048

Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>