[FIX] account_reports: filter translation
- Create several databases and install several languages
- Go to Accounts > Reports > Partner Ledger

The 'Account' filter might show 'Receivable' and 'Payable' in another
language than the one used by the user.

The corresponding `filter_account_type` variable is defined in the
class, and therefore translated at first initialization of the server.
Once initialized, it is reused globally and not evaluated based on the
user language.

The workaround is to redefine it when it is actually used, in the
`_build_options` method. We keep the original definition to avoid
breaking overrides expecting the field to exist without calling
`_build_options`.

Similar to 966b68179e44e42f169b8cccdcc695502c1b5bbc

Co-authored-by: Martin Trigaux <mat@odoo.com>
opw-1966941

closes odoo/enterprise#4074

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>