[FIX] web_studio: same field in 2 embed views=>ok
If we have eg. an added embedded tree and form view of a one2many:

```
<xpath expr="//field[name='field_one2many' position="inside">
  <tree>
    <field name="hello"/>
    <field name="world"/>
  </tree>
  <form>
    <field name="hello"/>
    <field name="everyone"/>
  </form>
</xpath>
```

After normalization, this would end up in:

```
<xpath expr="//field[name='field_one2many' position="inside">
  <tree>
    <field name="hello"/>
    <field name="everyone"/>
    <field name="world"/>
  </tree>
  <form>
    <field name="hello"/>
  </form>
</xpath>
```

Because when we have an opened new XPath with normalized content, we
check if the element before (or if none, the parent element) is in the
XPath, by searching the field in the current new XPath a match on:

- element tag name
- element attributes (name, class, and so on)
- element text

And in this instance, we have two such nodes.

With this changeset, we search the element going in the reverse order of
the depth-first iteration, so we get the last element matching what is
seeked.

The added test without change failed with `<div name="last_in_form"/>`
after `<div name="last_in_tree"/>` instead of original position.

opw-1950093
closes #4082

Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>