[FIX] website_enterprise: prevent multiple spinners on main nav-bar
Currently, if user clicks on 'Applications' icon on main nav-bar
of website to navigate to back-end, a spinner is added to indicate
initiation of the action. However, if user quickly clicks this icon
multiple times, more than one spinners are added for each click
(because event is being listened for every click, which also delays
the actual action).

This commit fixes the issue by listening to the event only
once when 'Applications' icon is clicked for the first time.

Task: 1963329
Closes: #4096

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>