[FIX] l10n_mx_edi_external_trade: remove calls to deprecated compute method
The compute and _compute method were deprecated in commit 71c56868.
They should be replaced by _convert, which takes explicit company and date
parameters, instead of taking them from the context.
An additional bug was added in these methods: they both simply use the _convert
method, but they forgot to pass the round parameter.
As a result, any call to compute is done with round=True,
even if explicitly given False.

We remove these calls, so that these methods can be removed entirely in master.

opw 1967738

closes odoo/enterprise#4098

Signed-off-by: Nans Lefebvre (len) <len@odoo.com>