[FIX] documents: upload to folder now working 12.0
In 12.0 088d918f0 if we upload from "Documents" app an attachment to a
folder not linked to a record, we would get an error with someone that
is not document manager because when creating attachment, we verify that
linked record:

 `self.env[attachment.res_model].browse(attachment.res_id)`

is writable. Since 088d918f0 an attachment uploaded, above case create
attachment with values:

 `{'res_id': FOLDER_ID, 'res_model': 'documents.folder'}`

instead of:

 `{'folder_id': FOLDER_ID}`

which is handled correctly.

With this changeset, if we create or update an attachment linked to a
documents.folder record, the link is replaced with `folder_id` data.

note: only for 12.0 since in 12.1 and over this is different and does
not happen.

opw-1964091
opw-1965832
opw-1967181
opw-1968063
closes #4102

Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>