[FIX] account_reports: bank reconciliation report with 'paid directly' vouchers
when using account_voucher module, users have the possiblity to record sales/expenses that are already paid. These create (in v11) one journal entry from the bank account to the income/expense in the sale/purchase journal + an account_payment in the bank journal.
Because the single entry for the move wasn't in the bank journal it simply wasn't shown in the bank reconciliation journal. v12 doesn't have the issue as the voucher split the voucher in 2 journal entries, one in each journal but in the meantime this little fix will allow the vouchers directly paid to appear in the report as expected.

Was OPW 1955317

closes odoo/enterprise#4105

Signed-off-by: Quentin De Paoli (qdp) <qdp@openerp.com>