[FIX] web_studio: report edit>breadcrumb>back OK
When going to report editor, then Report > More (visible in debug mode)
then going Back to "Report Editor" by breadcrumb: Boom there is a
traceback error in firefox or chromium.

This is because, detaching an iframe from DOM (when changing action)
then reattach it to DOM (going Back to "Report Editor") the iframe
content is wiped out, and the code does not expect it.

Example of the browser behavior:
 https://jsfiddle.net/fxbcdm4e/

In firefox and chromium, end result is:

```
before detach:[hello world]
when detached:no content window
after reattach:[]
```

but previous code expected `[hello world]` in `after reattach`.

With this changeset, we do not resize two times the iframe (it is
already done in any instance in the _updateContent that will set the
content back after it has been lost).

Without changeset, added test failed eg. on chrome with error:

```
  "TypeError: Cannot read property 'getComputedStyle' of null
     at getStyles (/web/static/lib/jquery/jquery.js:6118:55)
     ...
     at Class._resizeIframe (/web_studio/static/src/js/reports/report_editor.js:646:50)
     ...
     at Object.<anonymous> (/web_studio/static/tests/reports/report_editor_manager_tests.js:2148:23)
     ...
```

opw-1968093
closes #4121

Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>