[FIX] delivery_bpost: use the warehouse address to send shipment
The bpost_send_shipping method uses the warehouse_id.partner_id as sender.
To be coherent, this same address should be used consistently.

This might be a blocking problem if the company address is not in the same
country as the warehouse, as the address check fails,
making it imposible to validate a delivery.

opw 1966566

closes odoo/enterprise#4120

Signed-off-by: Nans Lefebvre (len) <len@odoo.com>