[ADD] sale_subscription_taxcloud: payment generation
1. Install sale_subscription, account_taxcloud and website_sale.
2. Set up Taxcloud
3. Create customer with an American address
4. Set "Fiscal Position for United States" to the customer.
5. Create subscription product.
   Do not put sales tax because tax will be coming from Taxcloud.
   The subscription template must have "Payment Mode" set to "Invoice
   only on successful payment"
6. Online payment acquirers should be activated. On the acquirer, set
   "Payment Flow" to "Payment from Odoo".
7. Purchase the subscription product in the eCommerce
9. To simulate repayment, set a recurring_next_date in the subscription
   in the past, then run the cron for the subscription.

A draft invoice will be generated and a payment is finalized. The
payment contains the base amount (without the tax), same as the invoice.

The usual Taxcloud flow adds the tax at invoice validation, not at
invoice creation. In this case, we create the payment, then potentially
validate the invoice => the payment doesn't include the tax.

No choice than adding a new module to add the tax computation before
creating the payment transaction...

opw-1970077

closes odoo/enterprise#4162

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>