[FIX] account_reports_followup, snailmail_account_reports_followup: do not crash on missing level in _execute_followup_partner
Basically a null_reference exception.
Copy-pasted code from get_followup_informations:
    followup_line = self.env['account_followup.followup.line'].browse(level[0])
except in that function, a check that level is non-null was made before.
This check was missing in both cases.

Because of this, clicking on "Process Follow-Ups" would result in a traceback.

Steps to reproduce the bug:
    - Have a customer C with invoice I1 overdue.
    - Process the follow-up (F1) on C. It is sent.
    - Have invoice I2 become overdue for C.
    - Try to make the followup (F2) on C. Traceback.

When F1 is done, payment_next_action_date is set to a certain date,
before which no follow-up should be done.
When I2 becomes overdue, C becomes 'in need of action'.
However get_followup_level returns False in that case,
because payment_next_action_date > current_date.

Because a mail (and potentially a paper-mail) could be sent each time
the user push the button, the behaviour of get_followup_level should be kept.

opw 1959534

closes odoo/enterprise#4185

Signed-off-by: Nans Lefebvre (len) <len@odoo.com>