[FIX] web_studio: report content sized as printed
The report editor show a report in the real size of chosen paperformat.

Eg. if we have a report with width=30mm and height=22mm, it will be
showed on Odoo in with=113px and height=83px (1mm == 3.78px)

So when printing, we have the same width/height size.

But the content is printed with a given DPI (meaning depth per inch,
which result in number of pixels per inch) whilst the report editor
just use 96 as DPI.

So if out paperformat has 200 DPI, browser DPI is 96 so the content will
be zoomed at 96/200 ratio when printing, but our preview does not take
DPI into account and display content at 1/1 ratio that does not match
what is printed.

With this changest, we apply on the previous a zoom of:

{PAPERFORMAT DPI / 96}

or a zoom to shrink content to fit page width as wkhtmltopdf do.

Do note that in firefox this currently do not work and the preview will
just be the same as before:
https://caniuse.com/#feat=css-zoom

Without the changeset, 6 of the added test assertions failed, the only
one that passed was "preview width same than paperformat".

ps: the box-sizing of the iframe container has been changed so iframe
has width of page minus horizontal paddings, and paddings are now used
as mm (that is what wkhtmltopdf do) instead of px.

opw-1934958
closes #4174

Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>