[FIX] helpdesk: send mail to customers from ticket created via kanban quick_create
If the partner_email is missing from the values at create,
the partner (given by its id) is not notified.
This is the case in the kanban quick_create view, not in the form view.
Note that in every case the partner_email was added on the record
by an explicit call to _onchange_partner_id.
However it happened after the ticket creation,
and thus if it not in the initial values the mail would not be sent.

opw 1964448

closes odoo/enterprise#4183

Signed-off-by: Nans Lefebvre (len) <len@odoo.com>