[FIX] sale_subscription: Subtotal on Subscription Lines only computed after Saving
Steps to reproduce the bug:
- Create a user A with subscription Manager access rights
- Create a subscription and add lines

Bug:

The subtotal was only computed after clicking on save

Technically, when adding a product on a subscription line, a new record in cache is created.
When making a sudo in fuction _compute_price_subtotal, the environment is changed
So the environments of variable line and sudo_line are not the same and checking
the pricelist with line and applying it on subtotal with sudo_line is not consistent
as this is not the same environment. In this case, the currency of the pricelist was defined
with line but not with sudo_line and then the price_subtotal was set to 0.0

Forward-port up to saas-11.3

opw:1962988

closes odoo/enterprise#4189

Signed-off-by: Simon Goffin (sig) <sig@openerp.com>