[FIX] stock_barcode: wait after saving form view
In the stock barcode client action, the use can open a classic form view.
This commit makes sure the tour completely leaves the form view before
sending again barcodes to interact with the client action.

This is due to the barcode listenner that is not stopped when enterring a product manualy

closes odoo/enterprise#4198

Signed-off-by: Simon Lejeune (sle) <sle@openerp.com>