[FIX] stock_barcode: double package transfer=>ok
If for example we do an internal transfer of a package form A to B then
from B to A with barcode, first transfer is OK but second fails with:

>  You cannot move the same package content more than once in the same
>  transfer or split the same package into two location.

This is because quant of a package are a view over "stock.quant" with
domain:
 ['|', ('quantity', '!=', 0), ('reserved_quantity', '!=', 0)]

On the first transfer, we only have one quant (with eg. 50 units) in
location A. On second transfer we have one quant in location A (with 0
unit) and a second one in location B (with eg. 50 units) and we would
have these two lines when scanning a package in a transfer.

With this changeset, the domain is used when getting quants of a package
so we get no empty quants in the picking.

opw-1964190
closes #4228

Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>