[FIX] account_asset: last day of purchase
- Create an asset with:
Depreciation Dates: Based on Last Day of Purchase Period
Prorata Temporis: true
Number of Depreciations: 12
Number of Months in a Period: 1
- Compute the depreciation

The depreciation day should be the last day of the month, each month.
However, it is not the case and ultimately becomes the 28th or 29th.

The use case is already taken into account for when Prorata Temporis is
false. There is no good reason to not do it in this case.

opw-1985328

closes odoo/enterprise#4273

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>