[FIX] quality_mrp_workorder: put back measure field
Commit f74712d8ec955e4e71a3406bc3c781eb118d9de0 remove field 'measure'
of workorder model from quality_mrp_workorder module to quality_mrp_iot.
The goal was to add a specific widget in case of iot suite installed.
The side effet was that the field 'measure' was missing when only
quality_mrp_workorder was installed.

This commit put back the missing field in the non-iot module and add
the specific widget in the iot module.

Issue : 1971035

closes odoo/enterprise#4279

Signed-off-by: Simon Lejeune (sle) <sle@openerp.com>