[FIX] account_invoice_extract: adaptation to new extract server version (120)
This commit adapt the code of account_invoice_extract to support new features from
the Extract server. Moreover, it reuses code from saas-12.1+ to keep only one version
more easy to maintain in the futur.

This new code fixs some wrong behaviours:

- If the server return some invoice lines, they will be created in the invoice if they are
coherent with the total and subtotal of the invoice
- If the server return some taxes information, it can be set on invoice lines, this will lead to
better invoice completion, as the total set in the invoice won't be the vat-included-total, but the
subtotal, so even if the tax is not detected, the invoice line should be correct, and the
user just have to set the tax on its line.
- As the code fix the invoice line behaviour, it gives too to the user the possibility to
choose between two behaviours: merge or not merge invoice lines into one invoice line (per detected tax)
As we are in stable, this option doesn't appear in the menu, but is available with a system parameter:
'account_invoice_extract.no_merging_lines_by_taxes'. The system merges taxes per default.
- The box renderer doesn't crash anymore when the invoice contains a lot of heavy pages.

closes odoo/enterprise#4269

Signed-off-by: pimodoo <pimodoo@users.noreply.github.com>