[FIX] account_reports: cash basis and analytic tags
- Activate cash basis
- Set a tax
Tax Due: Based on Payment
- Create an analytic account with a tag 'TEST'
- Create an invoice, add the tax and an analytic account on the invoice
line
- Validate and pay
- Open the P&L report
- Select 'Cash basis' and Analytic > Tags, filter on 'TEST'

The result is zero everywhere.

Any of the two filters work separately, but both at the same time don't
work.

The reason comes from the `WHERE` clause of the `payment_table`. It
includes the condition on the analytic tags, but the payment move lines
don't have the analytic account set. Therefore, no payment line is
found.

opw-1971751

closes odoo/enterprise#4288

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>