[FIX] account_reports: raise UserError with improper account configuration
To generate the Cash Flow statement, you need to have at least one bank
account configured to retrieve the moves.
Otherwise, a traceback is shown, not giving that information to the user,
only giving a ValueError/syntax error.

To avoid this vague traceback, we raise an UserError asking the user
to configure their Default Debit and Credit accounts
in the corresponding journals.

OPW-1972464

closes odoo/enterprise#4298

Signed-off-by: Nans Lefebvre (len) <len@odoo.com>