[FIX] helpdesk: only send the rating request to the customer
Partial backport of commit 257df50e.
When the rating request was sent to the customer, it was posted in the chatter.
As a result, all followers of the ticket would receive an email.
Starting from 257df50e it is sent as a note, and therefore only sent to the
customer.
It also make the behaviour more coherent with the rating on project tasks.

opw 1975201

closes odoo/enterprise#4301

Signed-off-by: Nans Lefebvre (len) <len@odoo.com>