[FIX] l10n_co_reports: don't overlap text in retention report headers
wkhtmltopdf does not seem to correctly handle Bootstrap, more
specifically it seems to have issues with display: flex.

In Bootstrap having multiple col-12's in one row results in each of
them being displayed on a separate row, i.e. in case of overflow the
elements wrap to the next line.

However when rendering with wkhtmltopdf this does not seem to be the
case. It seems the height of the row is restricted and either:

- the elements are on the same line and overlap each other, or
- only the first element shows

This restructures the header so it's rendered as originally intended
and so each given row adds up to 12 cols.

closes odoo/enterprise#4311

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>