[FIX] delivery_bpost: set timeout on bpost requests
When a request is made to the bpost servers, no timeout are made on the
request which lead to block the UI for 15 minutes which is the worker
timeout when bpost doesn't answer to the request.

To avoid this behaviour, we've added a timeout of 5 seconds to the
request and added a user error for the user.

We've also improved the description on label as it was taking
product_qty which always 0.

closes odoo/enterprise#4309

Signed-off-by: pimodoo <pimodoo@users.noreply.github.com>