[IMP] l10n_mx_edi: Support subscriptions with CFDI
The subscriptions model has the option to generate invoices in the
period, then the invoice is generated and after is send by email in the
same process:
https://github.com/odoo/enterprise/blob/12.0/sale_subscription/models/sale_subscription.py#L814

If the after_commit process is enabled, when is send the email, the PDF
file is not correctly generated, because the invoice is nos signed yet.
Then the customer receives the PDF incomplete, but in Odoo is correctly
signed the document.

With this fix, to avoid the depend to sale subscription, when are
executed the mexican process, verify that the field exists and that the
lines are related to one subscription. Then, in this case do not wait
for the next commit, and when the PDF is generated correctly.

closes odoo/enterprise#4320

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>