[FIX] l10n_mx_edi: disable after commit if there is a transaction payment
This fix tries to avoid the raise error `Some of the invoices that will
be paid with this record are not signed, and the UUID is required to
indicate the invoices that are paid with this CFDI`, when an invoice with
transactions payment is validated.

Currently, to link payments to invoice the invoice needs to be signed
but the signed service is called after commit to avoid sign an invoice
two times in case of errors. But to avoid no be able to validate
invoices from e-commerce sales we are disabling the after commit process
only for this kind of invoice.

Note: If the signing service fails, the raise error will be shown so
it'll be more difficult to know what kind of error the process has (this
needs to be improved in the master version).

closes odoo/enterprise#4347

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>